backend/requirements.txt
SQLAlchemy==2.0.4
pandas==1.5.3
fastapi==0.92.0
numpy==1.24.2
pydantic==1.10.5
python-dotenv==0.21.1
pytest==7.4.1
uvicorn==0.20
python-multipart==0.0.5
passlib==1.7.4
PyJWT==2.6.0
openpyxl==3.1.1
xlrd==2.0.1
starlette==0.25.0
python-dateutil==2.8.2
aiohttp==3.8.4
asyncpg==0.27.0
XlsxWriter==3.1.2
httpx==0.24.1
requests==2.28.2
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.3
anyio==3.7.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
et-xmlfile==1.1.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.5
py-make==0.1.1
pytest-dotenv==0.5.2
python-certifi-win32==1.6.1
urllib3==1.26.16
wrapt==1.16.0
yarl==1.9.2
